# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Engepe\u00e7as - Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/productDescriptions": {
            "get": {
                "tags": [
                    "ProductDescription"
                ],
                "summary": "Obtenha uma lista das descri\u00e7\u00f5es do produto.",
                "description": "Obtenha todas as descri\u00e7\u00f5es do produto ",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade limite de exibi\u00e7\u00e3o",
                        "required": false,
                        "type": "integer",
                        "default": "15"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "P\u00e1gina a ser exibida",
                        "required": false,
                        "type": "integer",
                        "default": "1"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "id"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Dire\u00e7\u00e3o da ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "ASC"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Informe a sele\u00e7\u00e3o de campos que devem retornar da consulta separados por virgula",
                        "required": false,
                        "type": "string",
                        "default": "id, hrd_D002_id, product_utilization_id, description, description_detail"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pesquise por qualquer campo, ao usar este campo as outras consultas ser\u00e3o desconsideradas",
                        "required": false,
                        "type": "string",
                        "default": "Busca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/ProductDescription"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProductDescription"
                ],
                "summary": "Armazene uma Descri\u00e7\u00e3o do Produto rec\u00e9m-criada no armazenamento.",
                "description": "Store ProductDescription",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Descri\u00e7\u00e3o do produto que deve ser armazenada.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ProductDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductDescription"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            }
        },
        "/productDescriptions/{id}": {
            "get": {
                "tags": [
                    "ProductDescription"
                ],
                "summary": "Exibir a descri\u00e7\u00e3o do produto especificada.",
                "description": "Obter Descri\u00e7\u00e3o do Produto ",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Produto Descri\u00e7\u00e3o.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductDescription"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ProductDescription"
                ],
                "summary": "Atualize a Descri\u00e7\u00e3o do Produto especificada no armazenamento.",
                "description": "Atualizar Descri\u00e7\u00e3o do Produto.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Descri\u00e7\u00e3o do Produto.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Descri\u00e7\u00e3o do produto que deve ser atualizada.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ProductDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/ProductDescription"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProductDescription"
                ],
                "summary": "Remova a Descri\u00e7\u00e3o Do Produto especificada do armazenamento ",
                "description": "Excluir Descri\u00e7\u00e3o do Produto ",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Descri\u00e7\u00e3o do Produto",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "ProductDescription": {
            "required": [
                "hrd_D002_id",
                "product_utilization_id",
                "description"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "hrd_D002_id": {
                    "description": "D002_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "product_utilization_id": {
                    "description": "D002_C008_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "description": {
                    "description": "D002_Descricao_Produto",
                    "type": "string"
                },
                "description_detail": {
                    "description": "D002_Decricao_Detalhada",
                    "type": "string"
                },
                "created_by": {
                    "description": "created_by",
                    "type": "integer",
                    "format": "int32"
                },
                "updated_by": {
                    "description": "updated_by",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_by": {
                    "description": "deleted_by",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "EngepecasAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}